Functions

LinearMPC.add_constraint! — Method

add_constraint!(mpc::MPC;
    Ax, Au, Ar, Aw, Ad, Aup,
    ub, lb, ks, soft, binary, prio)
add_constraint!(mpc;Ax,Au,ub,lb,
                ks, soft, binary,prio)

Adds the constraints lb ≤ Ax xₖ + Au uₖ ≤ ub for the time steps k ∈ ks (additional terms Ar rₖ, Aw wₖ, Ad dₖ, Aup u⁻ₖ are possible)

soft marks if the constraint should be softened (default false)

binary marks if either the upper or lower bounds should be enforced with equality (default false)

prio marks the relative priority of the constraint (default 0)

LinearMPC.certify — Method

certify(mpc; range, AS0, opts)

Provide certificates on the iteration complexity of DAQP for solving the resulting optimization problems.

range is the parameter range over which the certification should be done

AS0 is the starting working set in DAQP (defaults to empty)

settings the settings used in the certification (see ASCertain.CertSettings())

LinearMPC.compute_control — Method

compute_control(mpc,x;r,uprev)

For a given MPC mpc and state x, compute the optimal control action. Optional arguments:

r - reference value

uprev - previous control action

all of them defaults to zero.

LinearMPC.move_block! — Method

move_block!(mpc,block)

Reduce the number of controls by keeping it constant in blocks. For example, block=[2,1,3] keeps the control constant for 2 time-steps, 1 time step, and 3 time steps.

if sum(block) ≠ mpc.Nc, the resulting block will be padded or clipped

if block is an Int, a vector with constant block size is created

LinearMPC.mpc2mpqp — Method

mpc2mpqp(mpc)

For a given MPC structure mpc, form the multi-parametric QP mpQP.

LinearMPC.set_bounds! — Method

set_bounds!(mpc;umin,umax,ymin,umax)

Sets the bounds umin ≤ u ≤ umax and ymin ≤ y ≤ umax

LinearMPC.set_horizon! — Method

set_horizon!(mpc;Np,Nc)

Sets the prediction horizon Np and control horizon Nc

LinearMPC.set_input_bounds! — Method

set_input_bounds!(mpc;umin,umax)

Sets the input bounds umin ≤ u ≤ umax

LinearMPC.set_labels! — Method

set_labels!(mpc;x,u,y,d)

Sets the name of the states x, controls u, output u, disturbance d

LinearMPC.set_objective! — Method

set_objective!(mpc;Q,R,Rr,S,Qf)

Set the weights in the objective function `xN' C' Qf C xN^T + ∑ (C xₖ - rₖ)' Q (C xₖ - rₖ) + uₖ' R uₖ + Δuₖ' Rr Δuₖ + xₖ' S uₖ

A vector is interpreted as a diagonal matrix.

LinearMPC.set_output_bounds! — Method

set_output_bounds!(mpc;ymin,ymax,
                ks, soft, binary,prio)

Adds the constraints lb ≤ C x ≤ ub for the time steps k ∈ ks

soft marks if the constraint should be softened (default false)

binary marks if either the upper or lower bounds should be enforced with equality (default false)

prio marks the relative priority of the constraint (default 0)

LinearMPC.set_prestabilizing_feedback! — Method

set_prestabilizing_feedback!(mpc,K)

Sets the prestabilizing feedback K

LinearMPC.set_prestabilizing_feedback! — Method

set_prestabilizing_feedback!(mpc)

Sets the prestabilizing feedback K to the infinte horizon LQR gain`

LinearMPC.set_terminal_cost! — Method

set_terminal_cost!(mpc)

Sets the terminal cost Qf to the inifinite horizon LQR cost

LinearMPC.setup! — Method

setup!(mpc)

Sets up the mpc given its current parameters and settings Internally, this means generating an mpQP, and setting up a DAQP workspace.